refactor: arrange utils and fetch-handler
